refactor: queryTils to a single file
