--- packages/core/src/domains/posts/queryTils/query.ts
@@ -1,29 +1,0 @@
-import groq from 'groq';
-
-export const tilQuery = groq`
-*[_type == "til" && language in $languages && !(_id in path('drafts.**'))] | order(publishedAt desc){
-  _id,
-  publishedAt,
-  title,
-  language,
-  content[]{
-    ...,
-    markDefs[]{
-      ...,
-      _type == "internalLink" => {
-      ...,
-      "itemMeta": @.item -> {
-        "slug": slug.current,
-        _type
-      }
-    },
-    }
-  },
-  "slug": slug.current,
-  "tags": tags[]->{
-    _id,
-    name,
-    "slug": slug.current
-  }
-}
-`;
